+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added missing second `sourceName` parameter to `useSendMessage` closure.
+- Added `removeNested` to `Storage` and `StateInstance`.
+- React `StateProvider` will now call `removeNested` if it is the last provider
+  to listen to the state `stateNewData` event.
 
 ## Changed
 - The `stateNewData` event on `StateInstance` will now use `T` as the type